Fix: 각 모델의 id Datatype bigint => int 로 수정
구글 Oauth 로그인 시 구글이 제공하는 클라이언트의 uid 값이 매우 컸던 것으로
기억해서 bigint로 Datatype을 정했으나
해당 문제가 발생했을때 수정하는게 맞는 것 같아 bigint => integer 로 수정

+++ api-server/models/like.js
@@ -4,16 +4,16 @@ module.exports = (sequelize, DataTypes) => {
     {
       like_id: {
         allowNull: false,
-        type: DataTypes.BIGINT(11),
+        type: DataTypes.INTEGER,
         autoIncrement: true,
         primaryKey: true,
       },
       fk_user_id: {
-        type: DataTypes.BIGINT(11),
+        type: DataTypes.INTEGER,
         allowNull: false,
       },
       fk_video_id: {
-        type: DataTypes.BIGINT(11),
+        type: DataTypes.INTEGER,
         allowNull: false,
       },
     },
